+++ Backend/api/apps.py
@@ -1,0 +1,4 @@
+from django.apps import AppConfig
+class ApiConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'api'
