+++ Backend/api/urls.py
@@ -1,0 +1,8 @@
+# backend/api/urls.py
+from django.urls import path
+from .views import generate, poison
+
+urlpatterns = [
+    path('generate', generate, name='generate'),
+    path('poison', poison, name='poison'),
+]
